get the correct rate to 87%

+++ src/main/java/com/emotibot/correction/utils/SegementUtils.java
@@ -14,7 +14,18 @@ public class SegementUtils
 {
     public static void segementFile(String originFilePath, String targetFilePath) throws Exception
     {
-        boolean ret = checkFiles(originFilePath, targetFilePath);
+        segementFileNormal(originFilePath, targetFilePath);
+    }
+    
+    public static void segementFile(String originFilePath, String targetFilePath, String commonSentenceFilePath, String commonSentenceTargetFilePath) throws Exception
+    {
+        segementFileNormal(originFilePath, targetFilePath);
+        segementFileWithCommonSentence(commonSentenceFilePath, commonSentenceTargetFilePath);
+    }
+    
+    private static void segementFileNormal(String originFilePath, String targetFilePath) throws Exception
+    {
+        boolean ret = checkFiles(originFilePath, targetFilePath, true);
         if (!ret)
         {
             throw new Exception();
@@ -49,7 +60,47 @@ public class SegementUtils
         }
     }
     
-    private static boolean checkFiles(String originFilePath, String targetFilePath)
+    
+    private static void segementFileWithCommonSentence(String commonSentenceFilePath, String commonSentenceTargetFilePath) throws Exception
+    {
+        
+        boolean ret = checkFiles(commonSentenceFilePath, commonSentenceTargetFilePath, true);
+        if (!ret)
+        {
+            throw new Exception();
+        }
+        
+        BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream(commonSentenceFilePath)));
+        FileWriter fw = new FileWriter(commonSentenceTargetFilePath);
+        try
+        {
+            String line = null;
+            while((line = br.readLine()) != null)
+            {
+                List<Term> termList = StandardTokenizer.segment(line);
+                String output = "";
+                for(Term term : termList)
+                {
+                    output = output + term.word + " ";
+                }
+                fw.write(output.trim() + "\r\n");
+            }
+        }
+        finally
+        {
+            if (br != null)
+            {
+                br.close();
+            }
+            if (fw != null)
+            {
+                fw.close();
+            }
+        }
+    }
+    
+    
+    private static boolean checkFiles(String originFilePath, String targetFilePath, boolean rewrite)
     {
         File originFile = new File(originFilePath);
         if (!originFile.exists() || !originFile.isFile())
@@ -68,7 +119,10 @@ public class SegementUtils
             }
             else
             {
-                segementFile.delete();
+                if (rewrite)
+                {
+                    segementFile.delete();
+                }
             }
         }
         return true;
